requirements.txt
streamlit
sparknlp
PyPDF2
python-docx
pandas
pytest
pillow